Problem Set 4 Done

+++ java/src/test/java/edu/ahs/robotics/java/GridLoggerTest.java
@@ -60,7 +60,7 @@ public class GridLoggerTest {
         private boolean resetCalled = false;
 
         @Override
-        public double getCurrentTime(){
+        public long getCurrentTime(){
             return currentTime;
         }
